Updates Req of extending

+++ src/main/java/com/andy/controller/DonateController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
+import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -47,7 +48,7 @@ public class DonateController {
         try {
             donateService = donateDispatchService.getServiceByProvider(provider);
             RequestGatewayGreen aa = new RequestGatewayGreen();
-            aa.setAmount(1);
+            aa.setAmount(BigDecimal.valueOf(1));
             aa.setBrandName("Green");
             donateService.send(aa);
         } catch(ValidateException e) {
